+++ BinanceExchange.API/Websockets/AbstractBinanceWebSocketClient.cs
@@ -10,42 +10,38 @@ using BinanceExchange.API.Utility;
 using Newtonsoft.Json;
 using NLog;
 using WebSocketSharp;
-using Logger = NLog.Logger;
 
 namespace BinanceExchange.API.Websockets
 {
     /// <summary>
-    /// The BinanceWebSocketClient is used when wanting to open a connection to retrieve data through the WebSocket protocol. Implements IDisposable
+    /// Abstract class for creating WebSocketClients 
     /// </summary>
-    public class BinanceWebSocketClient : IDisposable, IBinanceWebSocketClient
+    public class AbstractBinanceWebSocketClient
     {
-        /// <summary>
-        /// These are provided to bypass an exception which occurs (down to the WebsocketSharp library)
-        /// </summary>
-        public SslProtocols SupportedProtocols { get; } = SslProtocols.Tls12 | SslProtocols.Tls11 | SslProtocols.Tls;
+        protected SslProtocols SupportedProtocols { get; } = SslProtocols.Tls12 | SslProtocols.Tls11 | SslProtocols.Tls;
 
         /// <summary>
         /// Base WebSocket URI for Binance API
         /// </summary>
-        private string _baseWebsocketUri = "wss://stream.binance.com:9443/ws";
+        protected string BaseWebsocketUri = "wss://stream.binance.com:9443/ws";
 
         /// <summary>
         /// Used for deletion on the fly
         /// </summary>
-        private Dictionary<Guid, BinanceWebSocket> _activeWebSockets;
-        private List<BinanceWebSocket> _allSockets;
-        private readonly IBinanceClient _binanceClient;
-        private ILogger _logger;
+        protected Dictionary<Guid, BinanceWebSocket> ActiveWebSockets;
+        protected List<BinanceWebSocket> AllSockets;
+        protected readonly IBinanceClient BinanceClient;
+        protected ILogger Logger;
 
-        private const string AccountEventType = "outboundAccountInfo";
-        private const string OrderTradeEventType = "executionReport";
+        protected const string AccountEventType = "outboundAccountInfo";
+        protected const string OrderTradeEventType = "executionReport";
 
-        public BinanceWebSocketClient(IBinanceClient binanceClient, ILogger logger = null)
+        public AbstractBinanceWebSocketClient(IBinanceClient binanceClient, ILogger logger = null)
         {
-            _binanceClient = binanceClient;
-            _activeWebSockets = new Dictionary<Guid, BinanceWebSocket>();
-            _allSockets = new List<BinanceWebSocket>();
-            _logger = logger ?? LogManager.GetCurrentClassLogger();
+            BinanceClient = binanceClient;
+            ActiveWebSockets = new Dictionary<Guid, BinanceWebSocket>();
+            AllSockets = new List<BinanceWebSocket>();
+            Logger = logger ?? LogManager.GetCurrentClassLogger();
         }
 
 
@@ -59,8 +55,8 @@ namespace BinanceExchange.API.Websockets
         public Guid ConnectToKlineWebSocket(string symbol, KlineInterval interval, BinanceWebSocketMessageHandler<BinanceKlineData> messageEventHandler)
         {
             Guard.AgainstNullOrEmpty(symbol, nameof(symbol));
-            _logger.Debug("Connecting to Kline Web Socket");
-            var endpoint = new Uri($"{_baseWebsocketUri}/{symbol.ToLower()}@kline_{EnumExtensions.GetEnumMemberValue(interval)}");
+            Logger.Debug("Connecting to Kline Web Socket");
+            var endpoint = new Uri($"{BaseWebsocketUri}/{symbol.ToLower()}@kline_{EnumExtensions.GetEnumMemberValue(interval)}");
             return CreateBinanceWebSocket(endpoint, messageEventHandler);
         }
 
@@ -73,8 +69,8 @@ namespace BinanceExchange.API.Websockets
         public Guid ConnectToDepthWebSocket(string symbol, BinanceWebSocketMessageHandler<BinanceDepthData> messageEventHandler)
         {
             Guard.AgainstNullOrEmpty(symbol, nameof(symbol));
-            _logger.Debug("Connecting to Depth Web Socket");
-            var endpoint = new Uri($"{_baseWebsocketUri}/{symbol.ToLower()}@depth");
+            Logger.Debug("Connecting to Depth Web Socket");
+            var endpoint = new Uri($"{BaseWebsocketUri}/{symbol.ToLower()}@depth");
             return CreateBinanceWebSocket(endpoint, messageEventHandler);
         }
 
@@ -87,8 +83,8 @@ namespace BinanceExchange.API.Websockets
         public Guid ConnectToTradesWebSocket(string symbol, BinanceWebSocketMessageHandler<BinanceAggregateTradeData> messageEventHandler)
         {
             Guard.AgainstNullOrEmpty(symbol, nameof(symbol));
-            _logger.Debug("Connecting to Trades Web Socket");
-            var endpoint = new Uri($"{_baseWebsocketUri}/{symbol.ToLower()}@aggTrades");
+            Logger.Debug("Connecting to Trades Web Socket");
+            var endpoint = new Uri($"{BaseWebsocketUri}/{symbol.ToLower()}@aggTrades");
             return CreateBinanceWebSocket(endpoint, messageEventHandler);
         }
 
@@ -99,11 +95,11 @@ namespace BinanceExchange.API.Websockets
         /// <returns></returns>
         public async Task<Guid> ConnectToUserDataWebSocket(UserDataWebSocketMessages userDataMessageHandlers)
         {
-            Guard.AgainstNull(_binanceClient, nameof(_binanceClient));
-            _logger.Debug("Connecting to User Data Web Socket");
-            var listenKey = await _binanceClient.StartUserDataStream();
+            Guard.AgainstNull(BinanceClient, nameof(BinanceClient));
+            Logger.Debug("Connecting to User Data Web Socket");
+            var listenKey = await BinanceClient.StartUserDataStream();
 
-            var endpoint = new Uri($"{_baseWebsocketUri}/{listenKey}");
+            var endpoint = new Uri($"{BaseWebsocketUri}/{listenKey}");
             return CreateUserDataBinanceWebSocket(endpoint, userDataMessageHandlers);
         }
 
@@ -112,11 +108,11 @@ namespace BinanceExchange.API.Websockets
             var websocket = new BinanceWebSocket(endpoint.AbsoluteUri);
             websocket.OnOpen += (sender, e) =>
             {
-                _logger.Debug($"WebSocket Opened:{endpoint.AbsoluteUri}");
+                Logger.Debug($"WebSocket Opened:{endpoint.AbsoluteUri}");
             };
             websocket.OnMessage += (sender, e) =>
             {
-                _logger.Debug($"WebSocket Message Received on Endpoint: {endpoint.AbsoluteUri}");
+                Logger.Debug($"WebSocket Message Received on Endpoint: {endpoint.AbsoluteUri}");
                 var primitive = JsonConvert.DeserializeObject<IWebSocketResponse>(e.Data);
                 switch (primitive.EventType)
                 {
@@ -141,7 +137,7 @@ namespace BinanceExchange.API.Websockets
             };
             websocket.OnError += (sender, e) =>
             {
-                _logger.Error($"WebSocket Error on {endpoint.AbsoluteUri}: ",e);
+                Logger.Error($"WebSocket Error on {endpoint.AbsoluteUri}: ", e);
                 CloseWebSocketInstance(websocket.Id, true);
                 throw new Exception("Binance UserData WebSocket failed")
                 {
@@ -151,8 +147,8 @@ namespace BinanceExchange.API.Websockets
                     }
                 };
             };
-            _activeWebSockets.TryAdd(websocket.Id, websocket);
-            _allSockets.Add(websocket);
+            ActiveWebSockets.TryAdd(websocket.Id, websocket);
+            AllSockets.Add(websocket);
             websocket.SslConfiguration.EnabledSslProtocols = SupportedProtocols;
             websocket.Connect();
 
@@ -164,18 +160,18 @@ namespace BinanceExchange.API.Websockets
             var websocket = new BinanceWebSocket(endpoint.AbsoluteUri);
             websocket.OnOpen += (sender, e) =>
             {
-                _logger.Debug($"WebSocket Opened:{endpoint.AbsoluteUri}");
+                Logger.Debug($"WebSocket Opened:{endpoint.AbsoluteUri}");
             };
             websocket.OnMessage += (sender, e) =>
             {
-                _logger.Debug($"WebSocket Messge Received on: {endpoint.AbsoluteUri}");
+                Logger.Debug($"WebSocket Messge Received on: {endpoint.AbsoluteUri}");
                 //TODO: Log message received
                 var data = JsonConvert.DeserializeObject<T>(e.Data);
                 messageEventHandler(data);
             };
             websocket.OnError += (sender, e) =>
             {
-                _logger.Debug($"WebSocket Error on {endpoint.AbsoluteUri}:", e);
+                Logger.Debug($"WebSocket Error on {endpoint.AbsoluteUri}:", e);
                 CloseWebSocketInstance(websocket.Id, true);
                 throw new Exception("Binance WebSocket failed")
                 {
@@ -185,8 +181,8 @@ namespace BinanceExchange.API.Websockets
                     }
                 };
             };
-            _activeWebSockets.TryAdd(websocket.Id, websocket);
-            _allSockets.Add(websocket);
+            ActiveWebSockets.TryAdd(websocket.Id, websocket);
+            AllSockets.Add(websocket);
             websocket.SslConfiguration.EnabledSslProtocols = SupportedProtocols;
             websocket.Connect();
 
@@ -200,10 +196,10 @@ namespace BinanceExchange.API.Websockets
         /// <param name="fromError"></param>
         public void CloseWebSocketInstance(Guid id, bool fromError = false)
         {
-            if (_activeWebSockets.ContainsKey(id))
+            if (ActiveWebSockets.ContainsKey(id))
             {
-                var ws = _activeWebSockets[id];
-                _activeWebSockets.Remove(id);
+                var ws = ActiveWebSockets[id];
+                ActiveWebSockets.Remove(id);
                 if (!fromError)
                 {
                     ws.CloseAsync(CloseStatusCode.PolicyViolation);
@@ -213,23 +209,6 @@ namespace BinanceExchange.API.Websockets
             {
                 throw new Exception($"No Websocket exists with the Id {id.ToString()}");
             }
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (!disposing) return;
-            _allSockets.ForEach(ws =>
-            {
-                if (ws.IsAlive) ws.Close(CloseStatusCode.Normal);
-            });
-            _allSockets = new List<BinanceWebSocket>();
-            _activeWebSockets = new Dictionary<Guid, BinanceWebSocket>();
         }
     }
 }